back/fastapi/doduk/requirements.txt
annotated-types==0.6.0
anyio
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httpx
idna==3.6
openai==1.6.1
pydantic
pydantic_core
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3>=1.26.0, <2.0.0
uvicorn==0.23.1
python-dotenv